requirements.txt
groq==0.9.0
python-dotenv==1.0.1
rich==13.9.2
openai==1.52.2
pydantic==2.9.2